pyproject.toml
[tool.poetry]
name = "mini-controlnet"
version = "0.1.0"
description = ""
authors = ["Vladislav Vinokurov <vinokurov.va@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
torch = "^2.8.0"
opencv-python = "^4.12.0.88"
torchvision = "^0.23.0"
pillow = "^11.3.0"
pre-commit = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
